requirements.txt
python-dotenv==0.21.0
requests==2.31.0
ruff==0.5.0
